feat(nx-aws-cdk): schematics
