+++ libs/aws-cdk-stack/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@flowaccount/aws-cdk-stack",
   "version": "99.99.99",
-  "schematics": "./generators.json",
+  
   "dependencies": {
     "@aws-cdk/core": "~1.121.0",
     "@aws-cdk/aws-ec2": "~1.121.0",
