--- libs/aws-cdk-stack/src/lib/schematics/application/files/environments/environment.staging.ts.template
@@ -1,133 +1,0 @@
-import { ServicePrincipal } from '@aws-cdk/aws-iam';
-import { Runtime } from '@aws-cdk/aws-lambda';
-import { Duration } from '@aws-cdk/core';
-import { IApplicationStackEnvironmentConfig } from '../../../../types';
-
-const _region = `<%= region %>`
-const _stage = `<%= stage %>`
-const _stackName = `<%= name %>`
-const _isProduction = `<%= isProduction %>`
-
-export const environment: IApplicationStackEnvironmentConfig = {
-  stackName: `${_stage}-${_app}`,
-  isProduction: _isProduction,
-  awsCredentials: { account: <%= accountid %>, region: <%= region %>},  
-  envName: _stage,
-  vpc: {
-    vpcAttributes: {
-        vpcId:  <%= vpcid %>,
-        availabilityZones: [ `${_region}a`, `${_region}b` ],
-        privateSubnetIds: [ <% for(let subnetId in vpcPrivateSubnetIds){ %> `<%= subnetId %>`, <% } %> ]
-    },
-    subnets: [ { <% for(let subnetId in subnets){ %> { subnetId: '<%= subnet.id %>', availabilityZone: '<%= subnet.availabilityZone %>' }, <% } %> ],
-  },
-  <% if(rds) { %>
-  aurora: {
-    securityGroupIds: [ <% for(let sgId in rdsSecurityGroupIds){ %> `<%= sgId %>`, <% } %> ],
-    isProduction: _isProduction,
-    username: '<%= rds.username %>', 
-    password: '<%= rds.password %>!' 
-  },
-  <% } %>
-  <% if(sqs) { %>
-  sqs: [
-    <% count=0;for(queue in sqs) { %>
-      {
-        queueName: `<%= queue.name %>`
-      
-    <% if(visibilityTimeout) { %>
-        visibilityTimeout: <%= queue.visibilityTimeout %>> // Duration.seconds() has to be same as the audit-trial-balance lambda consumer :L
-      },
-    <% } %>
-  ],
-  <% } %>
-    serverless: {
-      executionRole: {
-        name: `${_stage}-${_app}-lambda-execution-role`,
-        assumedBy: new ServicePrincipal(`lambda.amazonaws.com`),
-        policies: [
-            {
-              name: `${_stage}-${_app}-lambda-policy`,
-              statements: [
-                 {
-                    actions: [
-                      "logs:CreateLogGroup",
-                      "logs:CreateLogStream",
-                      "logs:PutLogEvents",
-                    ],
-                    resources: [ "*" ], // change this to be more secure
-                  },
-                  {
-                    actions: [
-                      "ec2:DescribeInstances",
-                      "ec2:CreateNetworkInterface",
-                      "ec2:AttachNetworkInterface",
-                      "ec2:DescribeNetworkInterfaces",
-                      "ec2:DeleteNetworkInterface",
-                      "autoscaling:CompleteLifecycleAction"
-                    ],
-                    resources: [ "*" ],  // change this to be more secure
-                  },
-                  <% if(sqs) { %>
-                  {
-                    forceResource: true,
-                    actions: [
-                      "sqs:DeleteMessage",
-                      "sqs:GetQueueUrl",
-                      "sqs:ChangeMessageVisibility",
-                      "sqs:ListDeadLetterSourceQueues",
-                      "sqs:SendMessageBatch",
-                      "sqs:ReceiveMessage",
-                      "sqs:SendMessage",
-                      "sqs:GetQueueAttributes",
-                      "sqs:ListQueueTags",
-                      "sqs:ChangeMessageVisibilityBatch"
-                    ], 
-                    resources: [ ],  // change this to be more secure
-                  }
-                  // you will have to add // "sqs:PurgeQueue",  "sqs:SetQueueAttributes", "sqs:SetQueueAttributes" by yourself (not recommended to use on production)
-                  <% } %>
-                ]
-              }
-            ]
-      },
-      <% if(lambdaFunctions) { %>
-      lambda: [
-        <% for(let lambdaFunction in lambdaFunctions){ %>
-            {
-              runtime: Runtime.NODEJS_14_X,
-              srcRootPath: '<%= root %>/src/app',
-              timeout: Duration.seconds(<%= lambdaFunction.timeout %>),
-              memorySize: <%= lambdaFunction.memorySize %>,
-              handler: "<%= lambdaFunction.handler %>",
-              <% if(lambdaFunction.securityGroupIds) { %>
-              securityGroupIds: [ <% for(let sgId in lambdaFunction.securityGroupIds){ %> `<%= sgId %>`, <% } %> ],
-              <% } %>
-              environment: {
-                ENV_NAME: `${this._stage}-${this._stackName}`,
-              },
-              functionName: `${_app}-<%= dasherize(lambdaFunction.name) %>`,
-              <% if(lambdaFunction.eventProperties && (lambdaFunction.eventProperties.kinesisEventSource || lambdaFunction.eventProperties.sqsEventSource)) { %>
-              eventProperties: {
-                <% if(lambdaFunction.eventProperties.kinesisEventSource) { %>
-                kinesisEventSource: {
-                  dataStreamArn: <%= lambdaFunction.eventProperties.kinesisEventSource.dataStreamArn %>,
-                }
-                <% } %>
-                <% if(lambdaFunction.eventProperties.sqsEventSource) { %>
-                  sqsEventSource: {
-                    properties: {
-                      batchSize: 10,
-                      maxBatchingWindow: 2,
-                      enabled: true
-                    }
-                  }
-                <% } %>
-              }
-              <% } %>
-            }
-        <% } %>
-        ]
-      <% } %>
-    }
-};
