--- libs/aws-cdk-stack/src/lib/schematics/application/files/src/main.ts.template
@@ -1,55 +1,0 @@
-// import { PreFirehoseStack, FireHoseStack, DbFirehoseStack } from './app';
-import { environment } from './environments/environment';
-import { logger } from '@nrwl/devkit';
-import { App } from '@aws-cdk/core';
-import { AuroraServerlessDbStack, ElasticsearchStack, ManagedPolicyStack, RoleStack, ServerlessApplicationStack, SQSStack, TypescriptLambdaStack, VpcStack } from '@flowaccount/aws-cdk-stack';
-
-logger.info("🐺")
-const app = new App();
-let resourceArns: string[] = []
-const _vpcStack = new VpcStack(app, `${environment.stackName}-vpc`, { ...environment.vpc, awsCredentials: environment.awsCredentials });
-<% if(sqs) { %>
-const sqsStack = new SQSStack(app,`${environment.stackName}-sqs`, { ...environment})
-resourceArns = resourceArns.concat(sqsStack.output.queues.map(q => q.queueArn));
-<% } %>
-// const esStack = new ElasticsearchStack(app, `${environment.stackName}-es`, {
-//   ...environment.elasticSearch,
-//   vpc: _vpcStack.vpc, 
-//   subnets: [ _vpcStack.subnetSelection ]
-// });
-<% if(rds) { %>
-const rdsStack = new AuroraServerlessDbStack(app, `${environment.stackName}-sls-db`, 
-{ vpc: _vpcStack.vpc, 
-  subnets: _vpcStack.subnetSelection,
-  ...environment.aurora
-});
-resourceArns.push(rdsStack.output.auroraArn)
-<% } %>
-// Initiate Serverless
-logger.debug(`start appending resources into policy`)
-          environment.serverless.executionRole.policies.forEach(policy => {
-            logger.debug(`start appending resources into statements`)
-            policy.statements?.forEach(statement => {
-              if(statement.forceResource){
-                resourceArns.forEach(arn => {
-                  logger.debug(`append ${arn} into policy`)
-                  statement.resources.push(arn)
-                })
-              }
-            })
-          })
-          const lambdaRoleStack = new RoleStack(app, `${environment.stackName}-serverless-role`, {
-            name: environment.serverless.executionRole.name,
-            assumedBy: environment.serverless.executionRole.assumedBy
-          })
-          environment.serverless.executionRole.policies.forEach(policyProps => {
-            logger.info(`creating stack ${policyProps.name}`)
-            const policy = (new ManagedPolicyStack(app, `${policyProps.name}`, { ...policyProps, roles: [ lambdaRoleStack.output.role ] })).output.policy
-          });
-          const lambdaStack = new TypescriptLambdaStack(app, `${environment.stackName}-serverless-typescript-lambda`, { 
-            vpc: _vpcStack.vpc,
-            subnets: _vpcStack.subnetSelection,
-            functions: environment.serverless.lambda,
-            role: lambdaRoleStack.output.role
-          });
-
