+++ libs/aws-cdk-stack/src/lib/stacks/index.ts
@@ -10,3 +10,4 @@ export * from './policy-stack';
 export * from './managed-policy-stack';
 export * from './dynamo-db-stack';
 export * from './rds-read-replica';
+export * from './builders'
