+++ libs/nx-aws-cdk/generators.json
@@ -1,17 +1,17 @@
 {
   "$schema": "../../node_modules/@angular-devkit/schematics/collection-schema.json",
-  "name": "@flowaccount/aws-cdk-stack",
+  "name": "@flowaccount/nx-aws-cdk",
   "generators": "./generators.json",
   "version": "0.0.1",
   "schematics": {
     "app-serverless": {
-      "factory": "./src/lib/schematics/application/app#appSchematic",
-      "schema": "./src/lib/schematics/application/schema.json",
+      "factory": "./src/schematics/application/app#appSchematic",
+      "schema": "./src/schematics/application/schema.json",
       "description": "Serverless application schematics, to create a serverless cdk stack on AWS"
     },
     "init": {
-      "factory": "./src/lib/schematics/init/init",
-      "schema": "./src/lib/schematics/init/schema.json",
+      "factory": "./src/schematics/init/init",
+      "schema": "./src/schematics/init/schema.json",
       "description": "nx aws-cdk initiation schematics. Installs aws-cdk to your repository",
       "aliases": ["init-aws-cdk"],
       "hidden": true
