+++ libs/nx-aws-cdk/src/schematics/application/files/.eslintrc.json
