+++ libs/nx-aws-cdk/src/schematics/application/files/src/app/handler.ts.template
