+++ libs/nx-aws-cdk/src/schematics/application/files/tsconfig.app.json
