+++ libs/nx-aws-cdk/src/schematics/application/schema.json
@@ -1,6 +1,5 @@
 {
   "$schema": "http://json-schema.org/schema",
-  "cli": "nx",
   "title": "Stack Options",
   "description": "Synth an app via the `cdk` cli command.",
   "type": "object",
@@ -20,6 +19,12 @@
       "default": "dev_console",
       "x-prompt": "What profile would you want to assume (profile from ~/.aws/config) ([profile <profile_name>) ?"
     },
+    "accountid": {
+      "type": "string",
+      "description": "Use the indicated AWS profile as the default environment",
+      "default": "xxxxxxxxxxx777",
+      "x-prompt": "What accountid would you want to deploy to ?"
+    },
     "region": {
       "type": "string",
       "description": "Use the indicated AWS profile as the default environment",
@@ -37,7 +42,6 @@
     "vpcId": {
       "type": "string",
       "description": "The vpc-id for this serverless application to live on",
-      "x-prompt": "What vpc id will this application be deployed to ?",
       "default": "vpc-"
     },
     "vpcPrivateSubnetIds": {
@@ -45,16 +49,14 @@
       "items": {
         "type": "string"
       },
-      "description": "Array of vpc private subnet ids",
-      "x-prompt": "Type in vpc private subnetids (minimum 1)? id1,id2,id3"
+      "description": "Array of vpc private subnet ids"
     },
-    "subnets": {
+    "subnetIds": {
       "type": "array",
       "items": {
         "type": "string"
       },
-      "description": "Array of lambda subnet ids",
-      "x-prompt": "What subnet ids will this application use? id1,id2,id3"
+      "description": "Array of lambda subnet ids"
     },
     "availabilityZones": {
       "type": "array",
@@ -62,12 +64,6 @@
       "uniqueItems": true,
       "items": {
         "type": "string"
-      },
-      "x-prompt": {
-        "message": "Which availability zones would you like to use?",
-        "type": "list",
-        "multiselect": true,
-        "items": ["a", "b", "c"]
       }
     },
     "handlers": {
@@ -75,16 +71,14 @@
       "items": {
         "type": "string"
       },
-      "description": "Array for lambda handler names",
-      "x-prompt": "What hadlers functions should be mapped to lambda functions (will map in linear order)? 'a','b','c' "
+      "description": "Array for lambda handler names"
     },
     "functionNames": {
       "type": "array",
       "items": {
         "type": "string"
       },
-      "description": "Array for lambda function names",
-      "x-prompt": "Function names for your lambda functions (will map in linear order)? 'a','b','c'"
+      "description": "Array for lambda function names"
     },
     "securityGroupIds": {
       "type": "array",
@@ -114,12 +108,6 @@
       "uniqueItems": true,
       "items": {
         "type": "string"
-      },
-      "x-prompt": {
-        "message": "Which event sources would you like to use?",
-        "type": "list",
-        "multiselect": true,
-        "items": ["kinesis", "sqs"]
       }
     },
     "eventsArn": {
@@ -131,14 +119,6 @@
     }
   },
   "required": [
-    "name",
-    "profile",
-    "region",
-    "vpcId",
-    "vpcPrivateSubnetIds",
-    "subnets",
-    "availabilityZones",
-    "handlers",
-    "functionNames"
+    "name"
   ]
 }
