+++ libs/nx-aws-cdk/src/schematics/base-aws-cdk.ts
