+++ libs/nx-aws-cdk/src/schematics/base-schema.d.ts
