+++ libs/nx-aws-cdk/src/schematics/common-schematics.ts
