+++ libs/nx-aws-cdk/src/schematics/ecs/files/app/tsconfig.cdk.json__tmpl__
