+++ libs/nx-aws-cdk/src/schematics/ecs/schema.d.ts
