+++ libs/nx-aws-cdk/src/schematics/ecs/schema.json
