+++ libs/nx-aws-cdk/src/schematics/init/init.spec.ts
