+++ libs/nx-aws-cdk/src/schematics/init/schema.d.ts
