+++ libs/nx-aws-cdk/src/schematics/normalized-schema.d.ts
