+++ libs/nx-aws-cdk/src/schematics/schema.d.ts
@@ -25,6 +25,7 @@ export function iscdkSchematicFlag(
 }
 
 export type Schema = {
+  accountid: string
   vpcPrivateSubnetIds: [];
   subnetIds: string[];
   availabilityZones: string[];
