+++ libs/nx-aws-cdk/src/schematics/schema.json
